v4: C++ migration docs, MAVLink C headers opt-in (#3097)
* docs: describe the C++ changes from v3 to v4

Adds a v4 section to the API changes page, covering what breaks when an
application moves from v3 to v4: the .hpp headers, building and linking
(C++ standard, CMake version check, soname, static dependencies, the
cpp/ source directory, dropped packages), the 32-bit system ID API, the
Telemetry, Shell, Info and Ftp plugin changes, the enums whose numeric
values changed, the MavlinkPassthrough deprecation, and behaviour
changes such as the abort when a System or plugin outlives its Mavsdk
and the new callback drop policy.

Also fixes guide pages that no longer matched: logging goes to stderr
and the header is log_callback.hpp, applications need C++17 rather
than C++11, and the user callback queue messages changed.

* core: keep the MAVLink C headers out of the public headers

Including mavsdk.hpp used to pull in the MAVLink C headers, so every
application got mavlink_message_t, MAV_* and friends in its namespace.
Only a few places in the public headers still needed them:

- The MAV_TYPE of a Configuration, and to_vehicle_from_mav_type(). These
  now use a new MavType enum, a copy of MAV_TYPE with the same values.
  mav_type.cpp checks every value against MAVLink at compile time, and
  fails when MAVLink gains a type that MavType does not have yet.
- The default component IDs, which are now spelled out as numbers and
  checked against MAV_COMP_ID_* in mavsdk.cpp.
- The deprecated MavlinkPassthrough plugin and
  intercept_incoming/outgoing_messages_async(), which are built on the
  MAVLink C types. They are now only available if
  MAVSDK_ENABLE_MAVLINK_C_API is defined, which also includes the MAVLink
  C headers again. Including mavlink_passthrough.hpp without it fails with
  an error pointing to MavlinkDirect.

The library, the system tests and the fuzzer define it for themselves,
so the library still contains everything it did before. Doxygen defines
it as well, so the API reference keeps documenting those APIs.

The C binding keeps its uint8_t MAV_TYPE and converts.

Also document the change in the v4 API changes, and add the Shell::Receive
reference page that #3042 did not commit.

* docs: add a MavlinkDirect quickstart and example links

The MavlinkDirect guide explained the pieces but never showed a program
that runs: its sections start in the middle with "// Connect ...". It
now opens with a complete program that connects, subscribes to
GLOBAL_POSITION_INT and prints what arrives, followed by the same
subscription parsing the JSON with nlohmann/json, including how to add
that dependency, since MAVSDK does not expose a JSON library in its
headers.

Both were compiled and run against a vehicle sending that message, so
the output shown is what they print.

The guide and the v4 API changes now also link the MavlinkDirect
examples in the repository, which is what someone deprecating their
MavlinkPassthrough usage is looking for.

Also editorial changes to the v4 API changes section.

* docs: link the parsing section instead of hinting in a comment

The commented-out nlohmann/json snippet in "Receiving messages" is now
covered properly by "Getting at the fields".

* examples: don't rely on the MAVLink C headers

The examples used MAVLink constants and <cstring> that they only got
because mavsdk.hpp included the MAVLink C headers.

The MavlinkDirect examples now spell out the values they use, which is
what a user of the runtime API has to do anyway, and events and
fly_multiple_drones include <cstring> for strcmp/strlen themselves.

Also apply clang-format to mav_type.cpp, which the style check flagged.

+++ cpp/src/mavsdk/core/mavsdk_impl.cpp
@@ -1330,7 +1330,8 @@ void MavsdkImpl::set_configuration_locked(Mavsdk::Configuration new_configuratio
         // that the previous default component will be deleted if it is not
         // used/referenced anywhere.
         _default_server_component = server_component_by_id_with_lock(
-            new_configuration.get_component_id(), new_configuration.get_mav_type());
+            new_configuration.get_component_id(),
+            static_cast<uint8_t>(new_configuration.get_mav_type()));
     }
 
     {
@@ -1340,7 +1341,7 @@ void MavsdkImpl::set_configuration_locked(Mavsdk::Configuration new_configuratio
     // We cache these values as atomic to avoid having to lock any mutex for them.
     _our_system_id = static_cast<uint8_t>(new_configuration.get_system_id());
     _our_component_id = new_configuration.get_component_id();
-    _our_mav_type = new_configuration.get_mav_type();
+    _our_mav_type = static_cast<uint8_t>(new_configuration.get_mav_type());
     _our_autopilot = new_configuration.get_autopilot();
     _our_compatibility_mode = new_configuration.get_compatibility_mode();
